requirements.txt
streamlit==1.54.0
plotly==6.5.2
pandas==2.2.0
resend==0.8.0
openai==1.12.0
python-dotenv==1.0.0
apscheduler==3.11.2
streamlit-aggrid==1.2.1.post2
pyyaml>=6.0
